QuantumCircuits

Documentation for QuantumCircuits.

QuantumCircuits.EastModelHamiltonian

QuantumCircuits.TFIMHamiltonian

QuantumCircuits.apply

QuantumCircuits.build_general_unitary_gate

QuantumCircuits.operation_tensor

QuantumCircuits.zero_state_tensor

QuantumCircuits.zero_state_vec

QuantumCircuits.EastModelHamiltonian — Type

A struct containing the parameters of the East Model.

The form of H represented by this struct is

\[\mathcal{H} = -e^{-s} \sqrt{c(1-c)} \left [ X_1 + \sum_{j=2}^N n_{j-1}X_j \right ] + (1-2c)\left [ n_1 + \sum_{j=2}^N n_{j-1}n_j + c \sum_{j=2}^N n_{j-1}\right ] + c\]

We precompute the scalars A and B such that

\[\begin{align*} A &= -e^{-s} \sqrt{c(1-c)} \\ B &= (1-2c) \end{align*}\]

QuantumCircuits.TFIMHamiltonian — Type

A struct containing the parameters of the Transverse Field Ising Model.

The form of H represented by this struct is

\[\mathcal{H} = -J \left ( \sum_{\langle i,j\rangle} \sigma_z^{(i)}\sigma_z^{(j)} + g \sum_i \sigma_x^{(i)} + h \sum_i \sigma_z^{(i)} \right ).\]

QuantumCircuits.apply — Method

Applies the given circuit to the state ψ.

The original state is not modified, unlike in the apply! version.

QuantumCircuits.build_general_unitary_gate — Method

Builds a two qubit gate based on rotation angles given in an array.

Must supply exactly 15 angles.

QuantumCircuits.operation_tensor — Method

operation_tensor(gates...)

Takes a list of gates to create a Kronecker product matrix that represents the operation.

QuantumCircuits.zero_state_tensor — Method

zero_state_vec([type=ComplexF64], n::Integer)

Creates a state tensor representing n qubits in the state |00...0>, with dimensions 2x2x...x2.

QuantumCircuits.zero_state_vec — Method

zero_state_vec([type=ComplexF64], n::Integer)

Creates a state vector representing n qubits in the state |00...0>.