Execute

QuantumCircuits.Execute.QiskitQuantum — Type

Qiskit Quantum Device

QuantumCircuits.Execute.QuantumSimulator — Type

Quantum Simulator

QuantumCircuits.Execute.execute — Function

Function execute the quantum circuit on simulator

QuantumCircuits.Execute.execute — Function

Function execute the quantum circuit on simulator

QuantumCircuits.Execute.getShift — Method

Function return the shift

QuantumCircuits.Execute.loss_expected_zero_state — Method

Loss method to check if final state is zero.

QuantumCircuits.Execute.qderivative — Function

Caclulate loss value and it derivateive

QuantumCircuits.Execute.qexecute — Function

Caclulate the the value.

QuantumCircuits.Execute.qjacobian — Function

Caclulate the jacobian on device.

QuantumCircuits.Execute.setAndConvert — Method

Set the parameters to the cicquit and convert to qiskit

QuantumCircuits.Execute.state2probability — Method

Method convert state to probability

QuantumCircuits.Execute.@bra_str — Macro

Macro proces ket binary vector to quantum state

QuantumCircuits.Execute.@ket_str — Macro

Macro proces ket binary vector to quantum state