QuantumCircuits.QCircuits.Circuit

Quantum Circuit

QuantumCircuits.QCircuits.Circuit.QCircuit — Type

QCircuit - Quantum Circuit structure

The fundamental element of quantum computing is the quantum circuit. A quantum circuit is a computational routine consisting of coherent quantum operations on quantum data, such as qubits. It is an ordered sequence of quantum gates, measurements and resets, which may be conditioned on real-time classical computation. A set of quantum gates is said to be universal if any unitary transformation of the quantum data can be efficiently approximated arbitrarily well as as sequence of gates in the set. Any quantum program can be represented by a sequence of quantum circuits and classical near-time computation.

In QuantumCircuits, this core element is represented by the QCircuit struct. Below is an example of a quantum circuit that makes a three-qubit GHZ state defined as:

\[|\psi\rangle = \frac{|000\rangle + |111\rangle}{\sqrt{2}}\]

using QuantumCircuits
qc = QCircuit(3)
qc.h(0)
qc.cx(0, 1)
qc.cx(1, 2)

      ┌───┐          
q0_0: ┤ H ├──■───────
      └───┘┌─┴─┐     
q0_1: ─────┤ X ├──■──
           └───┘┌─┴─┐
q0_2: ──────────┤ X ├
                └───┘
c0: 3/═══════════════
                     

Methods

Base.show — Method

Show method

QuantumCircuits.QCircuits.Circuit.getRandParameters — Method

Function create new random parameters vector

QuantumCircuits.QCircuits.Circuit.getparameters — Method

Function create new parameters vector

QuantumCircuits.QCircuits.Circuit.nop — Function

Nothing function

QuantumCircuits.QCircuits.Circuit.setClassicalRegister! — Method

Add the classical register to circuit.

QuantumCircuits.QCircuits.Circuit.toPythonQiskitCode — Method

Function to convert QCircuit to Python code using Qiskit library.

QuantumCircuits.QCircuits.Circuit.toQiskit — Method

Function to convert QCircuit to Qiskit circuit

QuantumCircuits.QCircuits.QBase.add! — Method

Add gate to circuit

QuantumCircuits.QCircuits.QBase.bindparameters! — Method

Function bind the parameters, so now there only values.

QuantumCircuits.QCircuits.QBase.measure! — Function

Add measures to circuit

QuantumCircuits.QCircuits.QBase.measure! — Function

Add measures to circuit

QuantumCircuits.QCircuits.QBase.setparameters! — Method

Save the parameters to the quantum circuits.

QuantumCircuits.QCircuits.Circuit.@addfunction — Macro

Add function macro