QuantumCircuits.QML.Optimization

Eva

QuantumCircuits.QML.Optimization.Eva — Type

Evolutionary Alpha optimization algorithm

The most popular optimization method in machine learning are is ADAM: "Adam: A Method for Stochastic Optimization" https://arxiv.org/abs/1412.6980. This method use only first derivative and the step is different for all of the parameters and it is changed in the way that combine the RMSprop and momentum method. I've found that better for your case work method which has different adaptive approach we calculate the estimation of the second derivative using the changes of derivative from two steps.

Methods

QuantumCircuits.QML.Optimization.Adam — Type

Adam optimization algoritm

QuantumCircuits.QML.Optimization.Momentum — Type

Momentum optimization algoritm

QuantumCircuits.QML.Optimization.RMSprop — Type

RMSprop optimization algoritm

QuantumCircuits.QML.Optimization.StandardGD — Type

Standard gradient descent