Random Matrices Generation

This module provides functions to generate random matrices.

Random Unitary

QuantumCircuits.haar_unitary — Function

haar_unitary(n::Int) :: Matrix{ComplexF64}

Generate a random Haar unitary matrix of size n x n.

Arguments

n::Int: The size of the matrix.

Returns

Matrix{ComplexF64}: The generated Haar unitary matrix.

using QuantumCircuits
haar_unitary(2)

2×2 Matrix{ComplexF64}:
 0.699742+0.443416im   0.451649-0.331295im
 0.403062+0.388953im  -0.752956+0.345419im

Random Hermitian

QuantumCircuits.random_hermitian — Function

random_hermitian(n::Int) :: Matrix{ComplexF64}

Generate a random Hermitian matrix of size n x n.

Arguments

n::Int: The size of the matrix.

Returns

Matrix{ComplexF64}: The generated Hermitian matrix.

using QuantumCircuits
random_hermitian(2)

2×2 Matrix{ComplexF64}:
  0.35615+0.0im      0.810639-2.01693im
 0.810639+2.01693im  0.293583+0.0im