Quantum Circuits Library: Single-qubit gates

U3

QuantumCircuits.QCircuits.Gates.U3 — Type

U3(qubit::Qubit, θ::Parameter, ϕ::Parameter, λ::Parameter)

Universal single-qubit rotation gate with three Euler angles, $\theta$, $\phi$ and $\lambda$. Matrix Representation

\[\newcommand{\th}{\frac{\theta}{2}} U3(\theta, \phi, \lambda) = \begin{pmatrix} \cos(\th) & -e^{i\lambda}\sin(\th) \\ e^{i\phi}\sin(\th) & e^{i(\phi+\lambda)}\cos(\th) \end{pmatrix}\]

qc = QCircuit(1)
qc.u3(0, 1, 2, 3)

You can also create a gate without adding parameters, in that case, they will be initialized by random values and can be used in Quantum Machine Learning as the loss function parameters.

using QuantumCircuits
using QuantumCircuits.QML
using QuantumCircuits.Execute

# Expected circuit
expqc = QCircuit(1)
expqc.x(0)
expmat = tomatrix(expqc)

# The ansact
qc = QCircuit(1)
qc.u3(0)

# Find the parameters which fit the expected unitary matrix in the best way.
findparam(expmat, qc)

Rx

QuantumCircuits.QCircuits.Gates.Rx — Type

Rx(qubit::Qubit, θ::Parameter)

A single-qubit Pauli gate which represents rotation about the X axis. Matrix Representation

\[\newcommand{\th}{\frac{\theta}{2}} Rx(\theta) = exp(-i \th X) = \begin{pmatrix} \cos{\th} & -i\sin{\th} \\ -i\sin{\th} & \cos{\th} \end{pmatrix}\]

qc = QCircuit(1)
qc.rx(0, π/2)

You can also create a gate without adding parameter, in that case, it will be initialized by random values and can be used in Quantum Machine Learning as the loss function parameters.

using QuantumCircuits
using QuantumCircuits.QML
using QuantumCircuits.Execute

# Expected circuit
expqc = QCircuit(1)
expqc.x(0)
expmat = tomatrix(expqc)

# The ansact
qc = QCircuit(1)
qc.rx(0)

# Find the parameters which fit the expected unitary matrix in the best way.
findparam(expmat, qc)

Ry

QuantumCircuits.QCircuits.Gates.Ry — Type

Ry(qubit::Qubit, θ::Parameter)

A single-qubit Pauli gate which represents rotation about the Y axis. Matrix Representation

\[\newcommand{\th}{\frac{\theta}{2}} Ry(\theta) = exp(-i \th Y) = \begin{pmatrix} \cos{\th} & -\sin{\th} \\ \sin{\th} & \cos{\th} \end{pmatrix}\]

qc = QCircuit(1)
qc.ry(0, π/2)

You can also create a gate without adding parameter, in that case, it will be initialized by random values and can be used in Quantum Machine Learning as the loss function parameters.

using QuantumCircuits
using QuantumCircuits.QML
using QuantumCircuits.Execute

# Expected circuit
expqc = QCircuit(1)
expqc.y(0)
expmat = tomatrix(expqc)

# The ansact
qc = QCircuit(1)
qc.ry(0)

# Find the parameters which fit the expected unitary matrix in the best way.
findparam(expmat, qc)

Rz

QuantumCircuits.QCircuits.Gates.Rz — Type

Rz(qubit::Qubit, θ::Parameter)

A single-qubit Pauli gate which represents rotation about the Z axis.

Matrix Representation

\[\newcommand{\th}{\frac{\theta}{2}} Rz(\theta) = exp(-i\th Z) = \begin{pmatrix} e^{-i\th} & 0 \\ 0 & e^{i\th} \end{pmatrix}\]

julia qc = QCircuit(1) qc.rz(0, π/2)

You can also create a gate without adding parameter, in that case, it will be initialized by random values and can be used in Quantum Machine Learning as the loss function parameters.

julia using QuantumCircuits using QuantumCircuits.QML using QuantumCircuits.Execute

Expected circuit

expqc = QCircuit(1) expqc.z(0) expmat = tomatrix(expqc)

The ansact

qc = QCircuit(1) qc.rz(0)

Find the parameters which fit the expected unitary matrix in the best way.

findparam(expmat, qc)

```

H

QuantumCircuits.QCircuits.Gates.H — Type

H(qubit::Qubit)

Single-qubit Hadamard gate, which is a $\pi$ rotation about the X+Z axis, thus equivalent to U3($\frac{\pi}{2},0,\pi$) Matrix Representation

\[H = \frac{1}{\sqrt{2}} \begin{pmatrix} 1 & 1 \\ 1 & -1 \end{pmatrix}\]

X

QuantumCircuits.QCircuits.Gates.X — Type

X(qubit::Qubit)

Single-qubit Pauli-X gate ($\sigma_x$), equivalent to U3($\pi,0,\pi$) Matrix Representation

\[X = \begin{pmatrix} 0 & 1 \\ 1 & 0 \end{pmatrix}\]

qc = QCircuit(1)
qc.x(0)

Y

QuantumCircuits.QCircuits.Gates.Y — Type

Y(qubit::Qubit)

Single-qubit Pauli-Y gate ($\sigma_y$), equivalent to U3($\pi,\frac{\pi}{2},\frac{\pi}{2}$) Matrix Representation

\[Y = \begin{pmatrix} 0 & -i \\ i & 0 \end{pmatrix}\]

qc = QCircuit(1)
qc.y(0)

Z

QuantumCircuits.QCircuits.Gates.Z — Type

Z(qubit::Qubit)

Single-qubit Pauli-Z gate ($\sigma_z$), equivalent to U3($0,0,\pi$) Matrix Representation

\[Z = \begin{pmatrix} 1 & 0 \\ 0 & -1 \end{pmatrix}\]

qc = QCircuit(1)
qc.z(0)

S

QuantumCircuits.QCircuits.Gates.S — Type

S(qubit::Qubit)

Single-qubit S gate, equivalent to U3($0,0,\frac{\pi}{2}$). This gate is also referred to a square-root of Pauli-Z. Matrix Representation

\[S = \begin{pmatrix} 1 & 0 \\ 0 & i \end{pmatrix}\]

qc = QCircuit(1)
qc.s(0)

Sd

QuantumCircuits.QCircuits.Gates.Sd — Type

Sd(qubit::Qubit) Single-qubit, hermitian conjugate of the S. This is also an alternative square root of the Z. Matrix Representation

\[S^{\dagger} = \begin{pmatrix} 1 & 0 \\ 0 & -i \end{pmatrix}\]

qc = QCircuit(1)
qc.sd(0)

T

QuantumCircuits.QCircuits.Gates.T — Type

T(qubit::Qubit)

Single-qubit T gate, equivalent to U3($0,0,\frac{\pi}{4}$). This gate is also referred to as a $\frac{\pi}{8}$ gate or as a fourth-root of Pauli-Z. Matrix Representation

\[T = \begin{pmatrix} 1 & 0 \\ 0 & e^{i\pi/4} \end{pmatrix}\]

qc = QCircuit(1)
qc.t(0)

Td

QuantumCircuits.QCircuits.Gates.Td — Type

Td(qubit::Qubit)

Single-qubit, hermitian conjugate of the T. This gate is equivalent to U3($0,0,-\frac{\pi}{4}$). This gate is also referred to as the fourth-root of Pauli-Z. Matrix Representation

\[T^{\dagger} = \begin{pmatrix} 1 & 0 \\ 0 & e^{-i\pi/4} \end{pmatrix}\]

qc = QCircuit(1)
qc.td(0)

Sx

QuantumCircuits.QCircuits.Gates.Sx — Type

Sx(qubit::Qubit)

Single-qubit square root of pauli-X. Matrix Representation

\[\sqrt{X} = \frac{1}{2} \begin{pmatrix} 1 + i & 1 - i \\ 1 - i & 1 + i \end{pmatrix}\]

qc = QCircuit(1)
qc.sx(0)

Sxd

QuantumCircuits.QCircuits.Gates.Sxd — Type

Sxd(qubit::Qubit)

Single-qubit hermitian conjugate of the square root of pauli-X, or the Sx. Matrix Representation

\[\sqrt{X}^{\dagger} = \frac{1}{2} \begin{pmatrix} 1 - i & 1 + i \\ 1 + i & 1 - i \end{pmatrix}\]

qc = QCircuit(1)
qc.sxd(0)